# src/main/resources/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "restfun 风格接口\n这里面的所有接口访问都需要权限验证\n都要在请求头中加入”authorization: Bearer “\n在空格后面加token\n用户登陆成功后返回的头中会有”authorization: Bearer token“",
    "version": "0.0.9-SNAPSHOT",
    "title": "铭贝api开发规范",
    "contact": {
      "name": "作者",
      "url": "https://baike.baidu.com/search/none?word=%E9%83%AD%E9%9B%A8%E8%A1%A1&pn=0&rn=10&enc=utf8",
      "email": "2350871838@qq.com"
    }
  },
...
  "host": "39.108.114.59",
  "basePath": "/",
...
  "paths": {
...
    "/system/dep": {
      "get": {
        "tags": [
          "单位"
        ],
        "summary": "获取该用户能修改的单位树",
        "operationId": "getDepsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Department"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "单位"
        ],
        "summary": "添加一个单位到另一个单位下",
        "operationId": "addDepUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "logo",
            "in": "formData",
            "description": "logo",
            "required": false,
            "type": "file"
          },
          {
            "name": "name",
            "in": "query",
            "description": "单位名称",
            "required": true,
            "type": "string"
          },
          {
            "name": "parentId",
            "in": "query",
            "description": "父节点id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/system/dep/{id}": {
      "put": {
        "tags": [
          "单位"
        ],
        "summary": "更新单位信息",
        "operationId": "updateDepUsingPUT",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "单位id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "logo",
            "in": "formData",
            "description": "logo",
            "required": false,
            "type": "file"
          },
          {
            "name": "name",
            "in": "query",
            "description": "单位名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "parentId",
            "in": "query",
            "description": "父节点id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "单位"
        ],
        "summary": "删除单位",
        "description": "删除前请确保不存在子单位且单位下不存在用户",
        "operationId": "deleteDepUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "单位id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/system/role": {
      "get": {
        "tags": [
          "系统角色"
        ],
        "summary": "获取所有的角色列表",
        "operationId": "getAllUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "系统角色"
        ],
        "summary": "添加一个角色",
        "operationId": "addRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "权限名称",
            "required": true,
            "type": "string"
          },
          {
            "name": "nameZh",
            "in": "query",
            "description": "角色中文名称（角色描述）",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/system/role/assignment/{id}": {
      "put": {
        "tags": [
          "系统角色"
        ],
        "summary": "为指定权限添加具体能访问的路径",
        "operationId": "addUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "角色id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "params",
            "description": "params",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/system/role/{id}": {
      "put": {
        "tags": [
          "系统角色"
        ],
        "summary": "更新角色信息",
        "operationId": "updateRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "角色id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "description": "角色名",
            "required": false,
            "type": "string"
          },
          {
            "name": "nameZh",
            "in": "query",
            "description": "权限的中文名称（权限描述）",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "系统角色"
        ],
        "summary": "删除一个角色",
        "operationId": "deleteRoleUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "Department": {
      "type": "object",
      "required": [
        "name",
        "parentId"
      ],
      "properties": {
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Department"
          }
        },
        "name": {
          "type": "string",
          "description": "单位名称"
        },
        "parentId": {
          "type": "integer",
          "format": "int64",
          "description": "父节点id"
        }
      },
      "title": "Department",
      "description": "单位"
    },
...
    "Role": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "description": "权限名称"
        },
        "nameZh": {
          "type": "string",
          "description": "权限的中文名称（权限描述）"
        }
      },
      "title": "Role"
    },
...
    "返回对象": {
      "type": "object",
      "properties": {
        "message": {
          "type": "object",
          "description": "对应信息"
        },
        "status": {
          "type": "string",
          "description": "状态码",
          "enum": [
            "200:操作成功",
            "400:操作失败（已知错误）",
            "304:操作失败（未知错误）"
          ]
        }
      },
      "title": "返回对象",
      "description": "统一返回格式"
    },
...
    "返回对象«string»": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "对应信息"
        },
        "status": {
          "type": "string",
          "description": "状态码",
          "enum": [
            "200:操作成功",
            "400:操作失败（已知错误）",
            "304:操作失败（未知错误）"
          ]
        }
      },
      "title": "返回对象«string»",
      "description": "统一返回格式"
    }
  }
}